chore: move activities components into admin folder
