+++ web/src/pages/Admin/ActivityPage/ActivityPage.js
@@ -1,4 +1,4 @@
-import ActivityCell from 'src/components/ActivityCell'
+import ActivityCell from 'src/components/Admin/ActivityCell'
 
 const ActivityPage = ({ id }) => {
   return <ActivityCell id={id} />
