+++ web/src/pages/Admin/EditActivityPage/EditActivityPage.js
@@ -1,4 +1,4 @@
-import EditActivityCell from 'src/components/EditActivityCell'
+import EditActivityCell from 'src/components/Admin/EditActivityCell'
 
 const EditActivityPage = ({ id }) => {
   return <EditActivityCell id={id} />
