+++ web/src/pages/Admin/NewActivityPage/NewActivityPage.js
@@ -1,4 +1,4 @@
-import NewActivity from 'src/components/NewActivity'
+import NewActivity from 'src/components/Admin/NewActivity'
 
 const NewActivityPage = () => {
   return <NewActivity />
